requirements.txt
anyio==3.6.2
async-generator==1.10
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
dnspython==2.3.0
et-xmlfile==1.1.0
exceptiongroup==1.1.1
fake-useragent==1.1.3
fonttools==4.39.3
greenlet==2.0.2
h11==0.14.0
htmlmin==0.1.12
httpcore==0.16.3
httpx==0.23.3
idna==3.4
ImageHash==4.3.1
importlib-resources==5.12.0
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
lxml==4.9.2
MarkupSafe==2.1.2
matplotlib==3.6.3
multimethod==1.9.1
networkx==3.1
numpy==1.23.5
openpyxl==3.1.2
outcome==1.2.0
packaging==23.1
pandas==1.5.3
pandas-profiling==3.6.6
patsy==0.5.3
phik==0.12.3
Pillow==9.5.0
pycparser==2.21
pydantic==1.10.7
pymongo==4.3.3
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-telegram-bot==20.2
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
requests==2.28.2
rfc3986==1.5.0
scikit-learn==1.2.2
scipy==1.9.3
scrapper-boilerplate==0.2.11
seaborn==0.12.2
selenium==4.9.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
SQLAlchemy==2.0.12
statsmodels==0.13.5
tangled-up-in-unicode==0.2.0
threadpoolctl==3.1.0
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.10.2
typeguard==2.13.3
typing_extensions==4.5.0
urllib3==1.26.15
visions==0.7.5
webdriver-manager==3.8.6
wsproto==1.2.0
ydata-profiling==4.1.2
zipp==3.15.0
